chore: add bottom padding to pages and organize component structure
- Add pb-36 to all page containers to prevent music player from covering content
- Organize components into feature-based folders (playlist, player, dashboard)
- Fix import paths after component reorganization
- Ensure all pages scroll properly with fixed music player at bottom
