+++ src/app/(protected)/likedtracks/page.tsx
@@ -43,7 +43,7 @@ const LikedTracks = () => {
   }, []);
 
   return (
-    <div className="p-6 max-w-7xl mx-auto bg-white min-h-screen">
+    <div className="p-6 max-w-7xl mx-auto bg-white min-h-screen pb-36">
       <h1 className="text-3xl font-bold text-gray-900 mb-8">
         Your Liked Tracks
       </h1>
