--- src/components/Footer.tsx
@@ -1,5 +1,0 @@
-const Footer = () => {
-  return <div>Footer</div>;
-};
-
-export default Footer;
